Upgrade to reveal.js-v4.1.0
To facilitate this I've had to change the way we do config---but I think
this new way is actually nicer anyway.

Basically, for configuration instead of having a revealjs-config.js file
you now specify the configuration in the config variable of a layout.

Links to dependency JavaScript files are similarly provided by a
`plugins` file.